pyproject.toml
[project]
name = "ridesapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "dotenv>=0.9.9",
    "fastapi>=0.135.2",
    "psycopg>=3.3.3",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.1",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
]
